wenku-assistant/requirements.txt
requests
fpdf
python-docx
openpyxl
flask
python-dotenv
re
json
time
os